+++ objectified-ui/public/WHATS_NEW.md
@@ -20,6 +20,7 @@ We continue to improve the platform based on your feedback with improvements and
   - Added the ability to test regular expressions directly within the form
   - Small improvements to the layouts and spacing for better readability
   - Editing a property of type "object" now gives visibility into its nested properties
+  - Added the ability to auto-generate an example value based on the property's schema
 - Improved status pill visibility with borders in light mode
 - Mermaid mode improvements:
   - Now adds display of the Mermaid diagram as an image by default